+++ src/common/document.ts
@@ -4,12 +4,16 @@ import { DocumentTypeFlag } from './flags';
 
 export const MAP_EXTENSIONS = ['.suma'];
 export const PROFILE_EXTENSIONS = ['.supr'];
+export const AST_EXTENSIONS = ['.ast.json'];
 
 export enum DocumentType {
   UNKNOWN = 'unknown',
   MAP = 'map',
   PROFILE = 'profile',
+  MAP_AST = 'map.ast',
+  PROFILE_AST = 'profile.ast',
 }
+
 /**
  * Detects whether the file on path is Superface Map or Superface Profile based on the extension.
  */
@@ -22,8 +26,20 @@ export function inferDocumentType(path: string): DocumentType {
     return DocumentType.PROFILE;
   }
 
+  for (const astExtension of AST_EXTENSIONS) {
+    if (MAP_EXTENSIONS.some(ex => normalizedPath.endsWith(ex + astExtension))) {
+      return DocumentType.MAP_AST;
+    }
+    if (
+      PROFILE_EXTENSIONS.some(ex => normalizedPath.endsWith(ex + astExtension))
+    ) {
+      return DocumentType.PROFILE_AST;
+    }
+  }
+
   return DocumentType.UNKNOWN;
 }
+
 /**
  * If flag is `DocumentTypeFlag.UNKNOWN` and `path` is defined, then calls `inferDocumentType(path)`
  * otherwise returns `flag`.
@@ -45,6 +61,7 @@ export function inferDocumentTypeWithFlag(
 
   return inferDocumentType(path);
 }
+
 export const DOCUMENT_PARSE_FUNCTION = {
   [DocumentType.MAP]: parseMap,
   [DocumentType.PROFILE]: parseProfile,